chore: doc: add copy-credits.py script
This script is used to export the credits from run.py to the credits
section in the README.md file. It is a simple script that reads the
credits from run.py and writes them to the README.md file.
It was written by me, with much help from the Microsoft Edge Copilot.
closes #53